feat: add abloom logo
