+++ src/modules/sections/home/orgSection/orgLogos.tsx
@@ -1,0 +1,179 @@
+const recruitingLogos = [
+  {
+    src: '/images/logos/recidiviz.png',
+    alt: 'Recidiviz',
+    url: 'https://www.recidiviz.org/',
+  },
+  {
+    src: '/images/logos/Humans_of_Public_Service.jpeg',
+    alt: 'Humans of Public Service Logo',
+    url: 'https://humansofpublicservice.org/about',
+  },
+  {
+    src: '/images/logos/Avela.png',
+    alt: 'Avela',
+    url: 'https://avela.org/',
+  },
+  {
+    src: '/images/logos/koko.png',
+    alt: 'Koko',
+    url: 'https://www.kokocares.org/',
+  },
+  {
+    src: '/images/logos/AmeelioLogo.png',
+    alt: 'Ameelio Logo',
+    url: 'https://www.ameelio.org/',
+  },
+  {
+    src: '/images/logos/Patrick_J_McGovern.png',
+    alt: 'Patrick J McGovern Logo',
+    url: 'https://www.mcgovern.org/',
+  },
+  {
+    src: '/images/logos/ello_logo.png',
+    alt: 'Ello',
+    url: 'https://www.helloello.com',
+  },
+  {
+    src: '/images/logos/McsilverLogo.png',
+    alt: 'NYU McSilver Logo',
+    url: 'https://www.mcsilver.nyu.edu/',
+  },
+  {
+    src: '/images/logos/New_Public_Logo.png',
+    alt: 'New_ Public',
+    url: 'https://newpublic.org/',
+  },
+  {
+    src: '/images/logos/Lead_For_America.jpg',
+    alt: 'Lead for America Logo',
+    url: 'https://www.leadforamerica.org/',
+  },
+  {
+    src: '/images/logos/new_data.png',
+    alt: 'New Data Logo',
+    url: 'https://center-for-new-data.breezy.hr/',
+  },
+  {
+    src: '/images/logos/GTSSE.png',
+    alt: 'Scientific Software Engineering Center at Georgia Tech',
+    url: 'https://ssecenter.cc.gatech.edu/',
+  },
+  {
+    src: '/images/logos/Project_Tech4Dev.png',
+    alt: 'Project Tech 4 Dev Logo',
+    url: 'https://projecttech4dev.org/',
+  },
+  {
+    src: '/images/logos/benetech.png',
+    alt: 'Benetech',
+    url: 'https://benetech.org/',
+  },
+  {
+    src: '/images/logos/schoolhouse.png',
+    alt: 'Schoolhouse',
+    url: 'https://schoolhouse.world/',
+  },
+  {
+    src: '/images/logos/Agency_Fund.png',
+    alt: 'Agency Fund Logo',
+    url: 'https://www.agency.fund/',
+  },
+  {
+    src: '/images/logos/Run_the_Future_logo.png',
+    alt: 'Run the Future',
+    url: 'https://formation.ventures/',
+  },
+  {
+    src: '/images/logos/FairOpportunityProject.png',
+    alt: 'Fair Opportunity Project',
+    url: 'https://www.fairopportunityproject.org/',
+  },
+  {
+    src: '/images/logos/giveinternet_logo.png',
+    alt: 'GiveInternet',
+    url: 'https://giveinternet.org/',
+  },
+  {
+    src: '/images/logos/I_CAN_HELP_HOST.png',
+    alt: 'I CAN HELP HOST',
+    url: 'https://icanhelp.host/',
+  },
+  {
+    src: '/images/logos/DataKind.png',
+    alt: 'DataKind',
+    url: 'https://www.datakind.org/',
+  },
+  {
+    src: '/images/logos/CommunityShare.png',
+    alt: 'CommunityShare',
+    url: 'https://www.communityshare.org/',
+  },
+  {
+    src: '/images/logos/Teaching_Lab.png',
+    alt: 'Teaching Lab',
+    url: 'https://www.teachinglab.org/',
+  },
+  {
+    src: '/images/logos/Uprooted.png',
+    alt: 'The Uprooted Way',
+    url: 'https://www.uprootedacademy.org/',
+  },
+  {
+    src: '/images/logos/ChangeMachine.png',
+    alt: 'Change Machine Logo',
+    url: 'https://change-machine.org/',
+  },
+  {
+    src: '/images/logos/ProtectDemocracy.png',
+    alt: 'Protect Democracy Logo',
+    url: 'https://protectdemocracy.org/',
+  },
+  {
+    src: '/images/logos/MurmurationLogo.png',
+    alt: 'Murmuration Logo',
+    url: 'https://murmuration.org/',
+  },
+  {
+    src: '/images/logos/PropelLogo.png',
+    alt: 'Propel Logo',
+    url: 'https://www.joinpropel.com/',
+  },
+  {
+    src: '/images/logos/1000More.png',
+    alt: '1000 More Logo',
+    url: 'https://www.1000more.org/',
+  },
+  {
+    src: '/images/logos/GovernForAmerica.png',
+    alt: 'Govern For America Logo',
+    url: 'https://www.govern4america.org/',
+  },
+  {
+    src: '/images/logos/CareerVillage.png',
+    alt: 'Career Village Logo',
+    url: 'https://www.careervillage.org/',
+  },
+  {
+    src: '/images/logos/Beam.png',
+    alt: 'Beam Logo',
+    url: 'https://www.bybeam.co/',
+  },
+  {
+    src: '/images/logos/SolarStewards.png',
+    alt: 'Solar Stewards Logo',
+    url: 'https://solarstewards.net/',
+  },
+  {
+    src: '/images/logos/CorticoLogo.png',
+    alt: 'Cortico Logo',
+    url: 'https://cortico.ai/',
+  },
+  {
+    src: '/images/logos/Abloom_logo.png',
+    alt: 'Abloom Logo',
+    url: 'https://abloomai.com/',
+  },
+];
+
+export default recruitingLogos;
